+++ CHANGES.md
@@ -1,2 +1,3 @@
 - Show enchanting scrolls as output item tooltip when enchanting a vellum
 - Fix executing some event handlers twice
+- Fix resourcefulness calculation